backend/requirements.txt
flask==2.0.3
Flask-CORS==3.0.10
mysql-connector-python==8.0.29
pandas==1.3.5
numpy==1.19.5
scikit-learn==1.0.2
tensorflow==2.7.0
matplotlib==3.5.1
seaborn==0.11.2
schedule==1.1.0
requests==2.28.2
python-dotenv==1.0.0
PyJWT==2.4.0
PyYAML==6.0.1